// src/enums/constants.ts
export enum ConstantsEnum {
  USERS = 'users',
  UPDATE_USER = 'update-user',
  EMAIL = 'email',
  VERIFY_CODE = 'verify-code',
  CODE = 'code',
  RESEND_CODE = 'resend-confirmation',
  CHANGE_PASSWORD = 'change-password',
  REDEFINE_PASSWORD = 'redefine-password',
  USER = 'User',
  SIGNIN = 'signin',
  SIGNUP = 'signup',
  GET_ALL = 'get-all',
  DELETE = 'delete',
  UPDATE = 'update',
  CONFIRM_CODE = 'confirm-code',
  ME = 'me',
  ID = '_id',
  Requester_ID = 'requesterId',
  PASSWORD = 'password',
  SAVE = 'save',
  JWT = 'jwt',
  CREATED_AT = 'createdAt',
  UPDATED_AT = 'updatedAt',
  CODE_CONFIRMATION = 'codeConfirmation',
  DDV = '__v',
  SERVER_IS_RUNNING = 'Server is running!',
  SUCESS_SEND_EMAIL = 'E-mail enviado com sucesso',
  NO_REPLY = 'Não Responda',
  CODE_OF_CONFIRMATION = 'Código de confirmação',
}